% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo.R
\name{combo}
\alias{combo}
\alias{wmst2}
\title{Comparing window mean survival time}
\usage{
combo(time, status, arm, type = NULL, combination = FALSE, round = FALSE, method = c("maximum", "average", "chi.sq")[1], tau0 = NULL, tau1 = NULL, alternative = c("two.sided", "greater", "less")[1], alpha = 0.05)
}
\arguments{
\item{time}{The follow-up time for right censored data.}

\item{status}{The status indicator, 1=event, and 0=right censored.}

\item{arm}{The group indicator for comparison. The elements of this vector take either 1 or 0. Normally, 0=control group, 1=active treatment group.}

\item{type}{The type of window selection method to be applied.}

\item{combination}{Determines whether the windows should be selected as every combination (TRUE) of the choices of tau0 and tau1 or not (FALSE).}

\item{round}{Determines whether the choices for tau0 and tau1 will be rounded to the nearest whole number (TRUE) or will remain as chosen (FALSE).}

\item{method}{Selects how the test statistics from each window will be combined.}

\item{tau0}{A scaler vector to specify the lower truncation time points of the window for the WMST calculation.
\code{tau0} needs to be at least 0 and smaller than tau1. When \code{tau0 = NULL}, the default value 0 is used (i.e., the lower bound for restricted mean survival analysis).}

\item{tau1}{A scaler vector to specify the upper truncation time points of the window for the WMST calculation.
\code{tau1} needs to be smaller than the minimum of the largest observed time in each of the two groups. When \code{tau1 = NULL}, the default value (i.e., the minimum of the largest observed time in each of the two groups) is used.}

\item{alternative}{Determines the sidedness of the alternative hypothesis. The default is "two.sided" which performs a two-tailed hypothesis test. If set to "greater" an upper-tailed test is performed. If set to "less" a lower-tailed test is performed.}

\item{alpha}{The default is 0.05. (1-\code{alpha}) confidence intervals are reported.}
}
\value{
an object of class combo.

\item{pvalue}{the pvalue produced from the versatile test of multiple windows.}

\item{tau}{the combination of various choices of tau0 and tau1 for windows.}
}
\description{
Performs two-sample comparisons using the window mean survival time (WMST) as a summary measure of the survival time distribution.
Three kinds of between-group contrast metrics (i.e., the difference in WMST, the ratio of WMST and the ratio of the window mean time lost (WMTL)) are computed.
The Greenwood plug-in estimator is used for the asymptotic variance.
}
\details{
For more details, please see the package vignette: \code{browseVignettes(package = "survWMST")}
}
\references{
Note to self: figure out what to place here!
}
\author{
Mitchell Paukner
}
