% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.wmst_pow.R
\name{print.pow}
\alias{print.pow}
\title{print.pow}
\usage{
\method{print}{pow}(x, digits = 5, ...)
}
\arguments{
\item{x}{Object to be printed.}

\item{digits}{Integer indicating the number of decimal places.}

\item{...}{Further arguments ignored in this function.}
}
\value{
returns summary output for class 'pow'
}
\description{
S3 method for class 'pow'
}
