% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.wmst.R
\name{print.wmst}
\alias{print.wmst}
\title{print.wmst}
\usage{
\method{print}{wmst}(x, digits = 5, ...)
}
\arguments{
\item{x}{Object to be printed.}

\item{digits}{Integer indicating the number of decimal places.}

\item{...}{Further arguments ignored in this function.}
}
\value{
returns summary output for class 'wmst'
}
\description{
S3 method for class 'wmst2'
}
