% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdata.R
\name{survdata}
\alias{survdata}
\title{Simulate Exponential Survival Data}
\usage{
survdata <- function(n, lambda1, lambda2, censor = 0, piecewise = FALSE, diverge_time = NULL, erate = NULL, eperiod = NULL, end_study = NULL)
}
\arguments{
\item{n0}{The sample size in the control arm.}

\item{n1}{The sample size in the treatment arm.}

\item{lambda1}{The rate of the exponential distribution in the first arm.}

\item{lambda2}{The rate of the exponential distribution in the second arm if piecewise = FALSE, or the rates of the piecewise exponential distribution of the second arm before diverge_time if piecewise = TRUE.}

\item{censor}{The probability a sample time will be deemed a censored observation.
\code{censor} needs to be between 0 and 1.}

\item{piecewise}{Determines if the second trial arm follows a piecewise exponential distribution.}

\item{diverge_time}{Determines the point in time in which the piecewise exponential curve changes rates.}

\item{erate}{Vector of enrollment rates following piecewise exponential distribution.}

\item{eperiod}{Vector of length +1 to erate which determines the times at which the enrollment period begins (0), when the erates change, and when the enrollment period ends.}

\item{end_study}{Creates a point at which all the simulated times are censored after.}
}
\value{
dataframe with observed times, calendar times, enrollment times, censoring indicators (0 for censoring, 1 for events), and trial arm.
}
\description{
Creates a data frame of survival data based on exponential distributions and Bernoulli censoring for a 2 arm trial.
}
\author{
Mitchell Paukner
}
