% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survplot.R
\name{survplot}
\alias{survplot}
\title{Plots two survival distributions.}
\usage{
survplot(survC = NA, survT = NA, xupper = NA)
}
\arguments{
\item{xupper}{the upper x axis limit for the plot.}

\item{survdefC}{the survival distribution of the control group (will be
plotted as a solid line), as a list in the form output by \code{survdef}.}

\item{survdefT}{the survival distribution of the control group (will be
plotted as a dashed line), as a list in the form output by \code{survdef}.}
}
\description{
Plots two survival distributions.
}
