% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survWM2-package.R
\docType{package}
\name{survWM2-package}
\alias{survWM2-package}
\title{Comparing Window Mean Survival Time}
\description{
Performs two-sample comparisons using the restricted mean survival time (RMST) as a summary measure of the survival time distribution.
Three kinds of between-group contrast metrics (i.e., the difference in WMST, the ratio of WMST and the ratio of the window mean time lost (WMTL)) are computed.
}
\examples{
#--- sample data ---#
D=rmst2.sample.data()
time=D$time
status=D$status
arm=D$arm
tau=NULL
x=D[,c(4,6,7)]
#--- without covariates ----
a=rmst2(time, status, arm, tau=10)
print(a)
plot(a, xlab="Years", ylab="Probability", density=60)
#--- with covariates ----
a=rmst2(time, status, arm, tau=10, covariates=x)
print(a)
}
\references{
Figure out references (might need to cite yourself)
}
\seealso{
survival
}
\author{
Mitchell Paukner

Maintainer: Mitchell Paukner \href{mailto:mpaukner@wisc.edu}{mpaukner@wisc.edu}
}
\keyword{survival}
