% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survWMST-package.R
\docType{package}
\name{survWMST-package}
\alias{survWMST-package}
\title{Comparing Window Mean Survival Time}
\description{
Performs two-sample comparisons using the window mean survival time (WMST) as a summary measure of the survival time distribution.
Three kinds of between-group contrast metrics (i.e., the difference in WMST, the ratio of WMST and the ratio of the window mean time lost (WMTL)) are computed.
Performs versatile tests for combinations of multiple windows.
}
\references{
Figure out references (might need to cite yourself)
}
\seealso{
survival
}
\author{
Mitchell Paukner

Maintainer: Mitchell Paukner \href{mailto:mpaukner@wisc.edu}{mpaukner@wisc.edu}
}
\keyword{survival}
