% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmst2.R
\name{wmst}
\alias{wmst}
\title{Comparing window mean survival time}
\usage{
rmst2(time, status, arm, tau0 = NULL, tau1 = NULL, alternative = "two.sided", alpha = 0.05)
}
\arguments{
\item{time}{The follow-up time for right censored data.}

\item{status}{The status indicator, 1=event, and 0=right censored.}

\item{arm}{The group indicator for comparison. The elements of this vector take either 1 or 0. Normally, 0=control group, 1=active treatment group.}

\item{tau0}{A scaler value to specify the lower truncation time point of the window for the WMST calculation.
\code{tau0} needs to be at least 0 and smaller than tau1. When \code{tau0 = NULL}, the default value 0 is used (i.e., the lower bound for restricted mean survival analysis).}

\item{tau1}{A scaler value to specify the upper truncation time point of the window for the WMST calculation.
\code{tau1} needs to be smaller than the minimum of the largest observed time in each of the two groups. When \code{tau1 = NULL}, the default value (i.e., the minimum of the largest observed time in each of the two groups) is used.}

\item{alternative}{Determines the sidedness of the alternative hypothesis. The default is "two.sided" which performs a two-tailed hypothesis test. If set to "greater" an upper-tailed test is performed. If set to "less" a lower-tailed test is performed.}

\item{alpha}{The default is 0.05. (1-\code{alpha}) confidence intervals are reported.}
}
\value{
an object of class wmst.

\item{tau0}{the lower truncation time used in the analyses}

\item{tau1}{the upper truncation time used in the analyses}

\item{note}{a note regarding the truncation time}

\item{WMST.arm1}{WMST results in arm 1.}

\item{WMST.arm0}{WMST results in arm 0.}

\item{result}{Results of the WMST analyses.}
}
\description{
Performs two-sample comparisons using the window mean survival time (WMST) as a summary measure of the survival time distribution.
Three kinds of between-group contrast metrics (i.e., the difference in WMST, the ratio of WMST and the ratio of the window mean time lost (WMTL)) are computed.
The Greenwood plug-in estimator is used for the asymptotic variance.
}
\details{
For more details, please see the package vignette: \code{browseVignettes(package = "survWMST")}
}
\examples{
#--- sample data ---#
D=wmst2_sample_data()
time=D$time
status=D$status
arm=D$arm
tau=NULL
x=D[,c(4,6,7)]
#--- without covariates ----
a=rmst2(time, status, arm, tau=10)
print(a)
plot(a, xlab="Years", ylab="Probability", density=60)
#--- with covariates ----
a=rmst2(time, status, arm, tau=10, covariates=x)
print(a)
}
\references{
Note to self: figure out what to place here! (Definitely give Hajime Uno, Lu Tian, Miki Horiguchi, Angel Cronin, Chakib Battioui, James Bell credit)
}
\author{
Mitchell Paukner
}
