% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmst_pow.R
\name{wmst_pow}
\alias{wmst_pow}
\title{Sample Size and Power for the Test of the Difference in Window Mean Survival Time}
\usage{
wmst_pow(survC, survT, censor = NA, eperiod, end_study, tau0, tau1, n = NA, power = NA, plot = FALSE, sim = F, M = 1000, method = "tau_star", alpha = 0.05, two.sided = FALSE)
}
\arguments{
\item{survC}{the survival distribution of the control group, as a list in the form output by \code{wmst_surv}.}

\item{survT}{the survival distribution of the treatment group, as a list in the form output by \code{wmst_surv}.}

\item{censor}{The common CDF of the censoring distribution for treatment and control arms.}

\item{eperiod}{The duration of the enrollment period if subjects are assumed to accrue uniformly over the enrollment interval \verb{(0,eperiod)}.}

\item{end_study}{The total duration of the study/follow-up period.}

\item{tau0}{A scaler vector to specify the lower truncation time points of the window for the WMST calculation.
\code{tau0} needs to be at least 0 and smaller than tau1. When \code{tau0 = NULL}, the default value 0 is used (i.e., the lower bound for restricted mean survival analysis).}

\item{tau1}{A scaler value to specify the upper truncation time point of the window for the WMST calculation.
\code{tau1} needs to be smaller than the minimum of the largest observed time in each of the two groups. When \code{tau1 = NULL}, the default value (i.e., the minimum of the largest observed time in each of the two groups) is used.}

\item{n}{The total sample size for both groups. 1:1 randomization is assumed.}

\item{power}{The desired power of the test.}

\item{plot}{If TRUE, plots of the assumed survival distribution and power as a function of sample size are produced.}

\item{alpha}{The type I error level. Default is 0.05.}

\item{two.sided}{Determines whether the test is upper-tailed (\code{two.sided} = FALSE) or two-sided. The default is FALSE.}
}
\value{
an object of class pow.

\item{n}{the user-specified n, or if n was left blank, the n needed to achieve the user-specified power.}

\item{powerWMST}{the user-specified power, or if power was left blank, the asymptotic power of the RMST test.
If \code{two.sided=FALSE}, \code{powerWMST} gives the power of rejecting in favor of the treatment arm.
If \code{two.sided=TRUE}, \code{powerWMST} gives the power of rejecting the null in favor of a two-sided alternative.}
}
\description{
Determine the asymptotic power of the test of WMST under a given trial design, or calculate the samples size needed to achieve a desired power.
}
\details{
For more details, please see the package vignette: \code{browseVignettes(package = "survWMST")}
}
\references{
Note to self: figure out what to place here!
}
\author{
Mitchell Paukner
}
