% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmst_surv.R
\name{wmst_surv}
\alias{wmst_surv}
\title{Create a Survdef Object for a Piecewise Exponential Distribution}
\usage{
wmst_surv(haz = NA, surv = NA, times = NA)
}
\arguments{
\item{haz}{a vector of hazards of length<=10. If a single hazard is
specified, the survival distribution is exponential with the specified
hazard. If \code{haz} has length>2, the survival distribution has constant hazard
equal to \code{haz} over the intervals \verb{[0, t_1), [t_1, t_2), \\dots, [t_m, Inf)}
where \code{t_i} are the entries of \code{times} and \code{times} has length \code{m}. One of
\code{haz}, \code{surv} must be specified.}

\item{surv}{a vector of survival probabilities of length<=10 corresponding to
\code{times}. If \code{surv} is specified, the survival distribution has constant hazard
over the intervals \verb{[0, t_1), [t_1, t_2), \\dots, [t_m, Inf)}
where \code{t_i} are the entries of \code{times} and \code{times} has length \code{m+1}. The hazards
are calculated so that the curve passes through each entry in \code{surv} at the
corresponding time from \code{times}. One of \code{haz}, \code{surv} must be specified.}

\item{times}{a vector of the same length as surv (if surv is specified) or
one element shorter than haz (if haz is specified). No times term is required
if a single hazard is specified in haz.}
}
\value{
a list with components:
\item{S}{a vectorized function that takes time as input and returns the survival probability at that time}
\item{h}{a vectorized function that takes time as input and returns the hazard at that time}
}
\description{
Create a new object which stores user-specified survival distribution
information in the format needed for the main function, \code{RMSTpow}. \code{survdef}
is used when the user wishes to specify a piecewise expoenential survival
distribution. Either the hazard on fixed intervals or survival probabilities
at fixed times can be specified.
}
