% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdata.R
\name{survdata}
\alias{survdata}
\title{Simulate Exponential Survival Data}
\usage{
rmst2(n, lambda1, lambda2, lambda3, censor = 0, piecewise = FALSE, diverge_time = NULL, end_study = NULL)
}
\arguments{
\item{n}{The sample size in each trial arm.}

\item{lambda1}{The rate of the exponential distribution in the first arm.}

\item{lambda2}{The rate of the exponential distribution in the second arm if piecewise = FALSE, or the rate of the exponential distribution of the second arm before diverge_time if piecewise = TRUE.}

\item{lambda3}{the rate of the exponential distribution of the second arm after diverge_time if piecewise = TRUE.}

\item{censor}{The probability a sample time will be deemed a censored observation.
\code{censor} needs to be between 0 and 1.}

\item{piecewise}{Determines if the second trial arm follows a piecewise exponential distribution.}

\item{diverge_time}{Determines the point in time in which the piecewise exponential curve changes rates from lambda 2 to lambda 3.}

\item{end_study}{Creates a point at which all the simulated times are censored after.}
}
\value{
dataframe with oberved times, censoring indicators (0 for censoring, 1 for events), and trial arm.
}
\description{
Creates a data frame of survival data based on exponential distributions and Bernoulli censoring for a 2 arm trial.
}
\references{
Note to self: figure out what to place here! (Definitely give Hajime Uno, Lu Tian, Miki Horiguchi, Angel Cronin, Chakib Battioui, James Bell credit)
}
\author{
Mitchell Paukner
}
